+++ sequencer/src/block2/queryable.rs
@@ -12,9 +12,9 @@ use super::{
     tx_iterator::{TxIndex, TxIterator},
 };
 
-impl QueryablePayload for Payload<u32> {
+impl QueryablePayload for Payload<u64> {
     type TransactionIndex = TxIndex;
-    type Iter<'a> = TxIterator<'a, u32>;
+    type Iter<'a> = TxIterator<'a, u64>;
     type InclusionProof = TxInclusionProof;
 
     fn len(&self, meta: &Self::Metadata) -> usize {
@@ -25,7 +25,7 @@ impl QueryablePayload for Payload<u32> {
         // (2) the number of ns table entries that could fit inside the ns table byte len
         // Why? Because (1) could be anything. A block should not be allowed to contain 4 billion 0-length nss.
         // The quantity (2) must exclude the prefix of the ns table because this prifix indicates only the length of the ns table, not an actual ns.
-        let ns_table = NameSpaceTable::<u32>::from_bytes(meta);
+        let ns_table = NameSpaceTable::<u64>::from_bytes(meta);
         let ns_table_len = ns_table.len();
 
         // First, collect the offsets of all the nss
